refactor: refactor `environment` into the `request`
