+++ curlz/src/curlz/request/environment/dot_env.rs
@@ -31,7 +31,7 @@ impl TryFrom<DotEnvFile> for Environment {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::workspace::tests::create_file;
+    use crate::test_utils::create_file;
     use indoc::indoc;
 
     #[test]
