+++ curlz/src/curlz/request/mod.rs
@@ -1,2 +1,4 @@
+pub mod backend;
 pub mod bookmark;
+pub mod environment;
 pub mod http;
